+++ stellar-contract-env-common/src/lib.rs
@@ -11,7 +11,7 @@ mod symbol;
 mod tagged_val;
 mod unimplemented_env;
 mod val;
-mod xdr_converter;
+mod obj_xdr_converter;
 
 // Re-export the XDR definitions
 pub use stellar_xdr as xdr;
@@ -29,7 +29,7 @@ pub use checked_env::CheckedEnv;
 pub use env::{Env, EnvBase};
 pub use env_val::{EnvVal, IntoEnvVal, IntoVal, TryFromVal, TryIntoEnvVal, TryIntoVal};
 pub use unimplemented_env::UnimplementedEnv;
-pub use xdr_converter::ObjectXdrConverter;
+pub use obj_xdr_converter::ObjectXdrConverter;
 
 // BitSet, Status and Symbol wrap RawVals.
 // TODO: maybe these should wrap EnvVals?
