+++ stellar-contract-env-common/src/obj_xdr_converter.rs
